% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcPredStats}
\alias{calcPredStats}
\title{Cacl prediciton stats}
\usage{
calcPredStats(df)
}
\arguments{
\item{df}{A data.frame}
}
\value{
A data.frame
}
\description{
\code{calcPredStats} calculates prediction stats related to AUC ROC
}
