% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{classif2Table}
\alias{classif2Table}
\title{Taxize classification to data frame}
\usage{
classif2Table(x, ranks)
}
\arguments{
\item{x}{Output from the `taxize::classification` function.}

\item{ranks}{Ranks to be selected. Default is given by `validRanks()`.}
}
\value{
A data frame.
}
\description{
\code{classif2Table} converts from classification format (`taxize` packatge)
to a data frame.
}
