% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterBSDB.R
\name{concatenateEvenSigs}
\alias{concatenateEvenSigs}
\title{Concatenate even signatures}
\usage{
concatenateEvenSigs(l, min_size = 5)
}
\arguments{
\item{l}{A list of signatures.}
}
\value{
A nested list of signatures by direction and condition (concatenated).
}
\description{
\code{concatenateEvenSigs} concatenate signatures that are the output
of the \code{filterEvenSignatures}.
}
