% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createDT}
\alias{createDT}
\title{Create DataTable}
\usage{
createDT(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
A DataTable output.
}
\description{
\code{craeteDT} creates a data table.
}
\details{
Reference for this code: https://www.r-bloggers.com/2019/12/vignette-downloadable-tables-in-rmarkdown-with-the-dt-package/
The link was valid on Dec 2, 2022.
}
