% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbbact.R
\name{dbEn2}
\alias{dbEn2}
\title{Run enrichment using the dbBact approach}
\usage{
dbEn2(control, case, term_list, freq = 1, perm = 1000)
}
\arguments{
\item{control}{Character vector. Down-regulated set (control/decreased).
Number of experiments must have been added as the "nexp" attribute to the
signature.}

\item{case}{Character vector. Up-regulated set (case/increased).
Number of experiments must have been added as the "nexp" attribute to the
signature.}

\item{term_list}{A list of signature terms (e.g., from bugphyzz).}

\item{freq}{Integer. Minimum frequency. Default is 1.}

\item{perm}{Number of permutations for calculating P-value with mean
difference.}
}
\description{
Run enrichment using the dbBact approach
}
