% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbbact.R
\name{dbMat}
\alias{dbMat}
\title{Create matrix for dbBact enrichment approach}
\usage{
dbMat(control, case, term_list, freq = 1)
}
\arguments{
\item{control}{Character vector. Down-regulated set (control/decreased).
Number of experiments must have been added as the "nexp" attribute to the
signature.}

\item{case}{Character vector. Up-regulated set (case/increased).
Number of experiments must have been added as the "nexp" attribute to the
signature.}

\item{term_list}{A list of signature terms (e.g., from bugphyzz).}

\item{freq}{Integer. Minimum frequency. Default is 1.}
}
\value{
A SummarizedExperiment.
}
\description{
\code{dbMat} creates a count matrix for control (decreased) and
case (incraesed) sets and signature terms (e.g., from bugphyzz).
}
