% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{.getTaxonClassification}
\alias{.getTaxonClassification}
\title{Get taxonomy classification}
\usage{
.getTaxonClassification(x)
}
\arguments{
\item{x}{A vector of taxonomy names or taxids.}
}
\value{
The classification of the first non-NA element of the input vector.
}
\description{
Get taxonomy classification
}
\keyword{internal}
