% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{elbows}
\alias{elbows}
\title{Elbows}
\usage{
elbows()
}
\value{
Vector of type double with thresholds.
}
\description{
\code{elbows} returns the prevalence threshold for TypicalMicrobiomeSignatures
based on the elbow of the curve method. See vignette.
}
\examples{

elbows()

}
