% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbbact.R
\name{filterDbEnRes}
\alias{filterDbEnRes}
\title{Filter Result of dbBact Enrichment}
\usage{
filterDbEnRes(se, min_val = 3, adj_pval = TRUE)
}
\arguments{
\item{se}{A SummarizedExperiment result of runnig dbBact enrichment.}

\item{min_val}{Minimun number of annotations in either case or control.
Integer.}

\item{adj_pval}{Logical. Whether adjust p-value or not. Default TRUE.}
}
\value{
A SummarizedExperiment.
}
\description{
\code{filterDbEnRes} filters result of dbBact
}
