% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterBSDB.R
\name{filterEvenDat}
\alias{filterEvenDat}
\title{Filter even dataset from BugSigDB}
\usage{
filterEvenDat(bsdb, min_size = 1)
}
\arguments{
\item{bsdb}{A BSDB data.frame.}

\item{min_size}{Number of signatures.}

\item{rank}{Character string "genus" or "species".}
}
\value{
A data.frame.
}
\description{
Filter even dataset from BugSigDB
}
