% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterTaxa}
\alias{filterTaxa}
\title{Filter taxa based on abundance values per sample}
\usage{
filterTaxa(x, min_ab = 1, min_per = 0.2)
}
\arguments{
\item{x}{A phyloseq or TreeSummarizedExperiment object with otu_table/assay
and sample_data/colData}

\item{min_ab}{The minimum value of abundance for taxon to be
considered as present in a sample. Default is 1. The default value of 1
could be good for counts. Relative abundance or other data
transformations might require another threshold value.}

\item{min_per}{minimum percentage of samples in which each taxon must be
present in order to be kept in the data. Default is 0.2. Taxon presence is
dtermined by the `min_ab` argument (see above).}
}
\value{
The filtered phyloseq/TreeSummarized object
}
\description{
\code{filterTaxa} filters the number of taxa per sample based on a minimum
value of abundance. This functions works with both phyloseq and
TreeSummarizedExperiment objects.
}
