% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attr_counts.R
\name{get_attr_counts}
\alias{get_attr_counts}
\title{Get counts for attribute}
\usage{
get_attr_counts(x)
}
\arguments{
\item{x}{Attribute dataframe imported from bugphyzz.}
}
\value{
A vector of counts per ranks.
}
\description{
\code{get_attr_counts} gets counts of attributes per rank.
}
