% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncbi_taxonomy.R
\name{get_ncbi_taxonomy}
\alias{get_ncbi_taxonomy}
\title{get_ncbi_taxonomy}
\usage{
get_ncbi_taxonomy(force_download = FALSE)
}
\arguments{
\item{force_download}{If FALSE (default) the function uses the ncbi
taxonomy in cache (if present). If TRUE the ncbi taxonomy is dowloaded.
If a taxonomy was present in the cache, it will be removed and replaced.}
}
\value{
A data frame with the complete NCBI taxonomy
}
\description{
\code{get_ncbi_taxonomy} downloads and imports the full taxonomic
classification from the NCBI database
}
