% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterBSDB.R
\name{getCatSignatures}
\alias{getCatSignatures}
\title{Concatenate signatures of a BugSigDB data.frame}
\usage{
getCatSignatures(bsdb, tax_level, min_size = 5)
}
\arguments{
\item{bsdb}{A BugSigDB data.frame.}

\item{tax_level}{Character string. "genus" or "species".}

\item{min_size}{Integer.}
}
\value{
Named nested list of signatures.
}
\description{
Concatenate signatures of a BugSigDB data.frame
}
