% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpaired.R
\name{getPairedSigs}
\alias{getPairedSigs}
\title{Get paired signatures}
\usage{
getPairedSigs(dec, inc, rank, cat = FALSE)
}
\arguments{
\item{dec}{Signatures decreased.}

\item{inc}{Signatures increased.}

\item{rank}{Taxonomic rank.}

\item{cat}{if TRUE concatenate signatures. Default is FALSE.}
}
\value{
A list.
}
\description{
Get paired signatures
}
