% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{getSets}
\alias{getSets}
\title{Get Sets}
\usage{
getSets(bsdb, bsdb_body_site, tms, tms_body_site, tax_level)
}
\arguments{
\item{bsdb}{Data imported from bsdb.}

\item{bsdb_body_site}{Body site in bsdb.}

\item{tms}{Tidy TypicalMicrobiomeSignatures imported with
\code{importTypicalTMS}}

\item{tms_body_site}{Body site in tms.}

\item{tax_level}{Taxonomic rank}
}
\value{
A list of bsdb sets and background sets (bsdb + tms).
}
\description{
\code{getSets} gets sets of signatures.
}
