% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_taxids}
\alias{import_taxids}
\title{Import taxids}
\usage{
import_taxids(x)
}
\arguments{
\item{x}{Name of the dataset from Microbiome BenchmarkData}
}
\value{
A data frame.
}
\description{
Import taxids
}
