% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{importBiosis}
\alias{importBiosis}
\title{Import biosis data}
\usage{
importBiosis()
}
\value{
A data.frame
}
\description{
\code{importBiosis} imports manually curated annotations used in the
paper of Calgaro and NYCHANES. It could be useful for comparison of
annotations
}
