% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{importNYCHANES}
\alias{importNYCHANES}
\title{Import the NYCHANES file}
\usage{
importNYCHANES()
}
\value{
A TreeSummarizedExperiment
}
\description{
\code{imporNYCHANES} imports the annotated NYCHANES file from the
original paper and repo. The output is already as a TreeSummarizedExperiment.
The imported TSE only contains data for 'Cigarette'  and 'Never Smokers'.
Some filtering might be necessary since many samples were removed from the
full study.
}
