% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{importTaxids}
\alias{importTaxids}
\title{Import taxids for MicrobiomeBenchmarkData}
\usage{
importTaxids(x = "HMP_2012_16S_gingival_V35_taxids")
}
\arguments{
\item{x}{Character string}
}
\value{
A data.frame.
}
\description{
\code{importTaxids} imports a data.frame for MicrobiomeBenchmarkData
}
