% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{importTidyTMS}
\alias{importTidyTMS}
\title{Import tidy Typical Microbiome Signatures}
\usage{
importTidyTMS(prevalence_threshold = 0.01)
}
\arguments{
\item{prevalence_threshold}{A numeric value: 0.01.}
}
\value{
A data.frame.
}
\description{
\code{importTidyTMS} imports prevalence data from the
TypicalMicrobiomeSignatures project.
}
