% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{importTMS}
\alias{importTMS}
\title{Import typical microbiome signatures}
\usage{
importTMS()
}
\value{
A data.frame.
}
\description{
\code{importTMS} Import all of the typical microbiome signatures in
a tidy format.
}
