% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{microbeSetEnrichment}
\alias{microbeSetEnrichment}
\title{Microbe Set Enrichment}
\usage{
microbeSetEnrichment(set, reference, sigs)
}
\arguments{
\item{set}{A character vector.}

\item{reference}{A character vector.}

\item{sigs}{A list of character vectors.}
}
\value{
A data.frame.
}
\description{
\code{microbeSetEnrichment} runs ORA (Fisher's exact test) comparing
a set of microbes (character vector), a reference (character vector),
and a list of signatures (list of vectors). Furthermore,
\code{microbeSetEnrichment} adjusts p values (FDR), and calculate odd ratios.
}
