% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{myDataTable}
\alias{myDataTable}
\title{Print datatable with my default options}
\usage{
myDataTable(dat, page_len = NULL)
}
\arguments{
\item{dat}{A data.frame.}

\item{page_len}{An integer. Number of rows to print.}
}
\value{
A datatable object
}
\description{
\code{myDataTable} prints a DT::datatable with my default options.
}
