% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{myDT}
\alias{myDT}
\title{My DataTable}
\usage{
myDT(df, cap = "Table. Caption...", ap = 0.1)
}
\arguments{
\item{df}{A data.frame.}

\item{cap}{A caption.}

\item{ap}{= ADJ.PVAL threshold.}
}
\value{
A data.table.
}
\description{
\code{myDT} prints a datatable
}
