% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{plotEnrichmentRes}
\alias{plotEnrichmentRes}
\title{Plot Enrichment Results}
\usage{
plotEnrichmentRes(res, body_site, tax_level)
}
\arguments{
\item{res}{Output from \code{runEnrichment}}

\item{body_site}{A character string for the title.}

\item{tax_level}{A character string for the title.}
}
\value{
A ggplot2 object.
}
\description{
\code{plotEnrichmentRes} plots the enrichment results.
}
