% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preSteps.R
\name{preSteps}
\alias{preSteps}
\title{Prepare data for propagation}
\usage{
preSteps(df, tax.id.type = "NCBI_ID", remove_false = TRUE)
}
\arguments{
\item{df}{A data.frame.}

\item{tax.id.type}{A character string. Default: 'NCBI_ID'.}

\item{remove_false}{TRUE or FALSE. Default TRUE.}
}
\value{
A data.frame.
}
\description{
\code{preSteps} prepares data for propagation.
}
