% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxNames2TaxIDs}
\alias{taxNames2TaxIDs}
\title{Taxonomic names to taxIDS}
\usage{
taxNames2TaxIDs(df, names_from)
}
\arguments{
\item{df}{A data frame.}

\item{names_from}{Name of the column with the names that should be used as
rownames. Default is to extract the rownames from the dataframe.}
}
\value{
A data frame with taxIDs instead of names.
}
\description{
Taxonomic names to taxIDS
}
