% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxTable2taxid}
\alias{taxTable2taxid}
\title{Convert taxonomic names to NCBI taxids}
\usage{
taxTable2taxid(df, names_from)
}
\arguments{
\item{df}{A data frame.}

\item{names_from}{Name of the column with the names that should be used as
rownames.}
}
\value{
A data frame with taxIDs instead of names.
}
\description{
\code{taxTable2Taxid} converts a table of taxonomic names to a table of
taxids. The format must be the same as curatedMetagenomicData (rowData) and
phyloseq (taxa_data).
}
\details{
This function should always be run in an interactive session because it will
ask for clarification when two taxonomic names belong to two different
taxids.
}
