% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classif2Table}
\alias{classif2Table}
\title{Taxize classification to data frame}
\usage{
classif2Table(x, ranks)
}
\arguments{
\item{x}{Output from the `taxize::classification` function.}

\item{ranks}{Ranks to be selected. Default is given by `valid_ranks()`.}
}
\value{
A data frame.
}
\description{
Taxize classification to data frame
}
