% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbbact.R
\name{dbEn2}
\alias{dbEn2}
\title{Create SE with dbBact approach}
\usage{
dbEn2(
  control,
  case,
  term_list,
  opt = "counts",
  f = "wilcox.test",
  prev = prev,
  perm = 1000
)
}
\arguments{
\item{control}{Set of controls (decreased)}

\item{case}{Set of cases (increased)}

\item{term_list}{List of sigs}

\item{opt}{"counts" or "scores"}
}
\value{
A SummarizedExperiment
}
\description{
Create SE with dbBact approach
}
