% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbbact.R
\name{dbHt}
\alias{dbHt}
\title{Plot heatmap of output of dbEn2}
\usage{
dbHt(se, row_pad = 2)
}
\arguments{
\item{se}{A SummarizedExperiment}
}
\value{
A Heatmap
}
\description{
Plot heatmap of output of dbEn2
}
