% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbbact.R
\name{dbMat}
\alias{dbMat}
\title{Create matrix for dbBact approach}
\usage{
dbMat(control, case, term_list, opt = "counts", prev = 0.25)
}
\arguments{
\item{term_list}{A list of signatures from bugphyzz.}

\item{opt}{Character string. "counts" (default) or "scores".}

\item{decreased}{Character vector. Down-regulated set (or control).}

\item{increased}{Character vector. Up-regulated set (or case).}
}
\value{
A matrix
}
\description{
\code{dbMat} creates a matrix of counts of control (decreased) and
case (increased) sets from BugSigDB.
}
