% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbbact.R
\name{dbMat}
\alias{dbMat}
\title{Create matrix for dbBact approach}
\usage{
dbMat(control, case, term_list, opt = "integer")
}
\arguments{
\item{term_list}{A list of signatures from bugphyzz.}

\item{opt}{Character string. Keywords accepted: "integer" or "double".
If "integer" (default), only presence and absence is considered. If
"double", the bugphyzz scores will be used (weighted mode). The latter
only works if the scores are included in the bugphyzz signatures in the
"Score" attribute.}

\item{decreased}{Character vector. Down-regulated set (or control).}

\item{increased}{Character vector. Up-regulated set (or case).}
}
\value{
A matrix
}
\description{
\code{dbMat} creates a matrix of counts of control (decreased) and
case (increased) sets from BugSigDB.
}
