% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{.getTaxonClassification}
\alias{.getTaxonClassification}
\title{Get taxonomy classification}
\usage{
.getTaxonClassification(x)
}
\arguments{
\item{x}{A vector of taxonomy names or taxids.}
}
\value{
A data frame with taxonomic classification (as in the taxize format).
}
\description{
\code{.getTaxonClassification} Gets the taxonomy clasification using
the `taxize:classification` function of the first non-NA element of an
input vector of taxids.
}
\keyword{internal}
